feat: repackage personal details and extract strings
