+++ feature/details/src/main/kotlin/com/divinelink/feature/details/person/ui/PersonalInfoSectionData.kt
@@ -1,0 +1,8 @@
+package com.divinelink.feature.details.person.ui
+
+import com.divinelink.core.ui.UIText
+
+data class PersonalInfoSectionData(
+  val title: UIText,
+  val value: UIText,
+)
